# openapi.yaml
openapi: 3.0.3
info:
  title: Meridian Control API
  version: 0.1.0
  description: |
    Internal control-tower API for a 3PL (warehouses, inventory, orders,
    last-mile shipments, fleet).

    ## Conventions (the single contract)

    **List envelope** — every collection endpoint returns exactly:
    `{ data, page, pageSize, total, sort, order }`

    **Error envelope** — every failure, including the 404 handler and the 429
    rate limiter:
    `{ error: { code, message, details? } }`

    **Types** — enums are lowercase; dates are ISO-8601 UTC strings; inventory
    quantities are integers; money is integer minor units; booleans are booleans.

    **Idempotency** — `POST /orders` and `POST /inventory/adjust` accept an
    `Idempotency-Key` header. A replay returns the stored response and sets
    `Idempotent-Replay: true`.
servers:
  - url: http://localhost:4000/api
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /inventory/levels:
    get:
      tags: [Inventory]
      summary: Inventory levels joined with product and warehouse
      parameters:
        - name: warehouseId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Page of inventory rows
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListEnvelope" }
  /inventory/adjust:
    post:
      tags: [Inventory]
      summary: Adjust on-hand quantity (stub — see NOTES.md)
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      responses:
        "501": { description: Not implemented in this timebox }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ListEnvelope:
      type: object
      description: The only list shape the API returns.
      required: [data, page, pageSize, total, sort, order]
      properties:
        data:
          type: array
          items: {}
        page: { type: integer, example: 1 }
        pageSize: { type: integer, example: 25 }
        total: { type: integer, example: 10000 }
        sort: { type: string, example: createdAt }
        order: { type: string, enum: [asc, desc], example: desc }
...